Made BIO_METHOD initialization for OpenSSL 1.1 thread-safe.
The implementation attempts to select one of the following implementations,
in the decreasing order of preference:

- On C++11 compilers, use function-local statics initialization, which is
  thread-safe.
- On POSIX systems, use pthreads as it supports static initialization.
- On all other systems, use PMutex from PTLib.

The last option is, strictly speaking, a regression from OpenSSL 1.0, which
allows to statically initialize the BIO_METHOD structure. It should not be
significant, though, as users are unlikely to work with PSSLChannel in global
constructors and destructors.