deadlock fix by Francisco Olarte:
When testing a functionality which created threads we found our
program was leaking memory, to the tune of aprox 12k RSS and exactly
260k vmData, untraceable by valgrind / ptlib memchecker.

We found the process acquired two new memory map for each cycle, a 4k
non accesible region, plus a 256k rw region, which exactly matched the
pattern for a thread stack ( guard page at the bottom of 256k stack ).

Looking through tlibthrd we found a suspicious conditional
pthread_detach, the only call, and no pthread_join anywhere, and
suspected zombie threads. Digging through the repos we found
https://sourceforge.net/p/opalvoip/code/28100/#diff-9 which, among
other things, makes detaching forced for every created thread, but trying
a simple backport of this lead to deadlocks.

A little investigation showed PThread destructor used m_thread id == 0
as a signal that the pthread had not been created, and so did not need
to be detached or taken out of the lists, but HouseKeepingThread::Main
was taking the thread out of the lists when autodeleting and setting
the id to 0 before deleting. So we just detached it manually before
deleting, and this seems to solve the problem.